Rename solver modules
This commit renames the specific solver packages solver modules to the
more generic name solver.  In other words the module containing the
solver definitions will always be in a module within a specific package
called solver.
