+++ src/petclaw/sharpclaw/solver.py
@@ -6,12 +6,12 @@ Module containing SharpClaw solvers for PetClaw
 import pyclaw.sharpclaw.sharpclaw
 
 
-class SharpClawSolver1D(pyclaw.sharpclaw.sharpclaw.SharpClawSolver1D):
+class SharpClawSolver1D(pyclaw.sharpclaw.solver.SharpClawSolver1D):
     """
     1D parallel SharpClaw solver.  See the corresponding PyClaw class for documentation.
     """
 
-class SharpClawSolver2D(pyclaw.sharpclaw.sharpclaw.SharpClawSolver2D):
+class SharpClawSolver2D(pyclaw.sharpclaw.solver.SharpClawSolver2D):
     """
     2D parallel SharpClaw solver.  See the corresponding PyClaw class for documentation.
     """
