+++ src/pyclaw/__init__.py
@@ -24,8 +24,8 @@ import riemann
 import plot
 
 __all__.extend(['ClawSolver1D','ClawSolver2D','ClawSolver3D','SharpClawSolver1D','SharpClawSolver2D'])
-from pyclaw.clawpack.clawpack import ClawSolver1D, ClawSolver2D, ClawSolver3D
-from pyclaw.sharpclaw.sharpclaw import SharpClawSolver1D, SharpClawSolver2D
+from pyclaw.clawpack.solver import ClawSolver1D, ClawSolver2D, ClawSolver3D
+from pyclaw.sharpclaw.solver import SharpClawSolver1D, SharpClawSolver2D
 
 
 # Sub-packages
